+++ stdb/index/seriesparser.h
@@ -29,223 +29,13 @@
 #include "stdb/common/basic.h"
 #include "stdb/common/status.h"
 #include "stdb/index/stringpool.h"
-#include "stdb/index/invertedindex.h"
+#include "stdb/index/series_matcher.h"
+#include "stdb/index/plain_series_matcher.h"
 
 namespace stdb {
 
-static const i64 STDB_STARTING_SERIES_ID = 1024;
-
-struct SeriesMatcherBase {
-  //! Series name descriptor - pointer to string, length, series id.
-  typedef std::tuple<const char*, int, i64> SeriesNameT;
-
-  virtual ~SeriesMatcherBase() = default;
-
-  /** Add new string to matcher.
-  */
-  virtual i64 add(const char* begin, const char* end) = 0;
-  virtual i64 add(const char* begin, const char* end, const Location& location) = 0;
-
-  /** Add value to matcher. This function should be
-   * used only to load data to matcher. Internal
-   * `series_id` counter wouldn't be affected by this call, so
-   * it should be set up propertly in constructor.
-   */
-  virtual void _add(std::string series, i64 id) = 0;
-
-  /** Add value to matcher. This function should be
-   * used only to load data to matcher. Internal
-   * `series_id` counter wouldn't be affected by this call, so
-   * it should be set up propertly in constructor.
-   */
-  virtual void _add(const char* begin, const char* end, i64 id) = 0;
-
-  /**
-   * Match string and return it's id. If string is new return 0.
-   */
-  virtual i64 match(const char* begin, const char* end) const = 0;
-
-  /**
-   * Convert id to string
-   */
-  virtual StringT id2str(i64 tokenid) const = 0;
-
-  /**
-   * pull new names
-   */
-  virtual void pull_new_names(std::vector<SeriesNameT>* buffer) = 0;
-
-  /**
-   * pull new names and locations
-   */
-  virtual void pull_new_names(std::vector<SeriesNameT>* name_buffer, std::vector<Location>* location_buffer) = 0;
-
-  /**
-   * get all ids
-   */
-  virtual std::vector<i64> get_all_ids() const = 0;
-};
-
-/** Series index. Can be used to retreive series names and ids by tags.
-  * Implements inverted index with compression and other optimizations.
-  * It's more efficient than PlainSeriesMatcher but it's costly to have
-  * many instances in one application.
-  */
-struct SeriesMatcher : SeriesMatcherBase {
-  //! Series name descriptor - pointer to string, length, series id.
-  // typedef std::tuple<const char*, int, i64> SeriesNameT;
-
-  typedef StringTools::TableT TableT;
-  typedef StringTools::InvT   InvT;
-
-  Index                    index;      //! Series name index and storage
-  TableT                   table;      //! Series table (name to id mapping)
-  InvT                     inv_table;  //! Ids table (id to name mapping)
-  i64                      series_id;  //! Series ID counter, positive values
-  //! are resurved for metrics, negative are for events
-  std::vector<SeriesNameT> names;      //! List of recently added names
-  std::vector<Location> locations;     //! List of recently added locations.
-  mutable std::mutex       mutex;      //! Mutex for shared data
-
-  SeriesMatcher(i64 starting_id = STDB_STARTING_SERIES_ID);
-
-  /** Add new string to matcher.
-  */
-  i64 add(const char* begin, const char* end) override;
-  i64 add(const char* begin, const char* end, const Location& location) override;
-
-  /** Add value to matcher. This function should be
-   * used only to load data to matcher. Internal
-   * `series_id` counter wouldn't be affected by this call, so
-   * it should be set up propertly in constructor.
-   */
-  void _add(std::string series, i64 id) override;
-
-  /** Add value to matcher. This function should be
-   * used only to load data to matcher. Internal
-   * `series_id` counter wouldn't be affected by this call, so
-   * it should be set up propertly in constructor.
-   */
-  void _add(const char* begin, const char* end, i64 id) override;
-
-  /**
-   * Match string and return it's id. If string is new return 0.
-   */
-  i64 match(const char* begin, const char* end) const override;
-
-  /**
-   * Convert id to string
-   */
-  StringT id2str(i64 tokenid) const override;
-
-  /** Push all new elements to the buffer.
-   * @param buffer is an output parameter that will receive new elements
-   */
-  void pull_new_names(std::vector<SeriesNameT>* buffer) override;
-
-  /**
-   * pull new names and locations to the buffer
-   */
-  void pull_new_names(std::vector<SeriesNameT>* name_buffer, std::vector<Location>* location_buffer) override;
-  
-  /**
-   * get all the ids
-   */
-  std::vector<i64> get_all_ids() const override;
-
-  std::vector<SeriesNameT> search(IndexQueryNodeBase const& query) const;
-
-  std::vector<StringT> suggest_metric(std::string prefix) const;
-
-  std::vector<StringT> suggest_tags(std::string metric, std::string tag_prefix) const;
-
-  std::vector<StringT> suggest_tag_values(std::string metric, std::string tag, std::string value_prefix) const;
-
-  size_t memory_use() const;
-
-  size_t index_memory_use() const;
-
-  size_t pool_memory_use() const;
-
- protected:
-  i64 add_impl(const char* begin, const char* end);
-};
-
-/** Series matcher. Table that maps series names to series
-  * ids.
-  * Implements simple forward index. Can be used to map ids
-  * to names and names to ids. Can search index using regular
-  * expressions.
-  */
-struct PlainSeriesMatcher : SeriesMatcherBase {
-  //! Series name descriptor - pointer to string, length, series id.
-  // typedef std::tuple<const char*, int, i64> SeriesNameT;
-
-  typedef StringTools::TableT TableT;
-  typedef StringTools::InvT   InvT;
-
-  // Variables
-  LegacyStringPool         pool;       //! String pool that stores time-series
-  TableT                   table;      //! Series table (name to id mapping)
-  InvT                     inv_table;  //! Ids table (id to name mapping)
-  i64                      series_id;  //! Series ID counter
-  std::vector<SeriesNameT> names;      //! List of recently added names
-  std::vector<Location> locations;     //! List of recently added locations
-  mutable std::mutex       mutex;      //! Mutex for shared data
-
-  PlainSeriesMatcher(i64 starting_id = STDB_STARTING_SERIES_ID);
-
-  /** Add new string to matcher.
-   */
-  i64 add(const char* begin, const char* end) override;
-  i64 add(const char* begin, const char* end, const Location& location) override;
-
-  /** Add value to matcher. This function should be
-   * used only to load data to matcher. Internal
-   * `series_id` counter wouldn't be affected by this call, so
-   * it should be set up propertly in constructor.
-   */
-  void _add(std::string series, i64 id) override;
-
-  /** Add value to matcher. This function should be
-   * used only to load data to matcher. Internal
-   * `series_id` counter wouldn't be affected by this call, so
-   * it should be set up propertly in constructor.
-   */
-  void _add(const char*  begin, const char* end, i64 id) override;
-
-  /** Match string and return it's id. If string is new return 0.
-  */
-  i64 match(const char* begin, const char* end) const override;
-
-  //! Convert id to string
-  StringT id2str(i64 tokenid) const override;
-
-  /** Push all new elements to the buffer.
-   * @param buffer is an output parameter that will receive new elements
-   */
-  void pull_new_names(std::vector<SeriesNameT>* buffer) override;
-
-  /**
-   * pull all new names and locations to the buffer
-   */
-  void pull_new_names(std::vector<SeriesNameT>* name_buffer, std::vector<Location>* location_buffer) override;
-
-  /**
-   * get all the ids
-   */
-  std::vector<i64> get_all_ids() const override;
-
-  std::vector<SeriesNameT> regex_match(const char* rexp) const;
-
-  std::vector<SeriesNameT> regex_match(const char* rexp, StringPoolOffset* offset, size_t* prevsize) const;
-
- protected:
-  i64 add_impl(const char* begin, const char* end);
-};
-
 /** Namespace class to store all parsing related things.
-  */
+ */
 struct SeriesParser {
   /** Convert input string to normal form.
    * In normal form metric name is followed by the list of key
